pyproject.toml
[tool.poetry]
name = "NJUlogin"
version = "0.0.0"
description = "The Nanjing University login module, which can be used to login to the various campus web sites"
authors = ["Do1e <dpj.email@qq.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Do1e/NJUlogin"
repository = "https://github.com/Do1e/NJUlogin"
packages = [
    {include = "NJUlogin"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
requests = "^2.32.3"
pillow = "^11.0.0"
numpy = "^2.1.3"
lxml = "^5.3.0"
pycryptodome = "^3.21.0"
user-agents = "^2.2.0"
onnxruntime = "^1.20.0"
cryptography = "^43.0.3"

[[tool.poetry.source]]
name = "njumirror"
url = "https://mirror.nju.edu.cn/pypi/web/simple"
priority = "primary"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pytest = "^8.3.3"
poetry-dynamic-versioning = {extras = ["plugin"], version = "^1.4.1"}

[tool.poetry-dynamic-versioning]
enable = true

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
